src/main/java/br/com/jairfreitas/Restapi/resources/exceptions/ResourceExceptionHandler.java
package br.com.jairfreitas.Restapi.resources.exceptions;

import br.com.jairfreitas.Restapi.services.exceptons.DataIntegrityViolationException;
import br.com.jairfreitas.Restapi.services.exceptons.ResourceNotFoundException;
import jakarta.servlet.http.HttpServletRequest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

import java.time.Instant;

@ControllerAdvice
public class ResourceExceptionHandler {

    @ExceptionHandler(ResourceNotFoundException.class)
    public ResponseEntity<StandardErro> resourceNotFound(ResourceNotFoundException ex, HttpServletRequest request){
        StandardErro error = new StandardErro(Instant.now(), HttpStatus.NOT_FOUND.value(), ex.getMessage(), request.getRequestURI());
        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(error);
    }
    @ExceptionHandler(DataIntegrityViolationException.class)
    public ResponseEntity<StandardErro> dataIntegrityViolationException(DataIntegrityViolationException ex, HttpServletRequest request){
        StandardErro error = new StandardErro(Instant.now(), HttpStatus.BAD_REQUEST.value(), ex.getMessage(), request.getRequestURI());
        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
    }
}
